forecast-admin/forecast/requirements.txt
Django==1.8.4
django-localflavor==1.1
djangorestframework==3.2.3

psycopg2==2.6.1

whitenoise==1.0.6  #manages static assets
dj-database-url==0.3.0 #grabs environment variables and dumps them into a Django settings file
waitress==0.8.9 #a pure python WSGI server that is a replacement for gunicorn